requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Deprecated==1.2.14
dj-database-url==2.2.0
Django==4.2
django-filter==24.3
django-model-utils==4.5.1
django-notifications-hq==1.8.3
django-phone-verify==3.0.0
django-phonenumber-field==8.0.0
django-taggit==6.0.0
djangorestframework==3.15.2
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
jsonfield==3.1.0
Markdown==3.6
multidict==6.0.5
nexmo==2.5.2
packaging==24.1
phonenumbers==8.13.43
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
python-usernames==1.0.0
pytz==2024.1
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
swapper==1.3.0
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
wrapt==1.16.0
yarl==1.9.4
